epai3-capstone/backend/requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
black==21.7b0
python-multipart==0.0.5
python-jose==3.3.0
cryptography==3.4.8
passlib==1.7.4
bcrypt==3.2.0